image_gen/requirements.txt

requests
python-dotenv
fal-client==0.7.0
